# perf: dont clone witness when proving