master:
```
== Running model metrics: elapsed=60.449222458s peak=1.5 GiB in_use=1.5 GiB ==
== Checking accuracy metrics: elapsed=4.625µs peak=1.5 GiB in_use=1.5 GiB ==
== Witness poly fields generation metrics elapsed=45.727657791s peak=9.1 GiB in_use=9.1 GiB ==
== Witness table multiplicities metrics elapsed=392.636875ms peak=9.2 GiB in_use=9.2 GiB ==
== Challenge storage metrics elapsed=17.625µs peak=9.2 GiB in_use=9.2 GiB ==
== Witness context metrics elapsed=46.121288792s peak=9.2 GiB in_use=9.2 GiB ==
== Claims generation metrics elapsed=88.943723208s peak=13.9 GiB in_use=9.3 GiB ==
== Generate proof metrics elapsed=6.155216709s peak=11.2 GiB in_use=1.9 GiB ==
== Proving metrics: elapsed=141.3274805s peak=1.9 GiB in_use=378.7 MiB ==
```
this branch:
```
== Running model metrics: elapsed=44.844812417s peak=1.5 GiB in_use=1.5 GiB ==
== Checking accuracy metrics: elapsed=4.333µs peak=1.5 GiB in_use=1.5 GiB ==
== Witness poly fields generation metrics elapsed=50.1269505s peak=9.1 GiB in_use=9.1 GiB ==
== Witness table multiplicities metrics elapsed=455.876083ms peak=9.2 GiB in_use=9.2 GiB ==
== Challenge storage metrics elapsed=20.625µs peak=9.2 GiB in_use=9.2 GiB ==
== Witness context metrics elapsed=50.583881917s peak=9.2 GiB in_use=9.2 GiB ==
== Claims generation metrics elapsed=84.513571916s peak=11.3 GiB in_use=9.3 GiB ==
== Generate proof metrics elapsed=7.525828083s peak=11.2 GiB in_use=1.9 GiB ==
== Proving metrics: elapsed=142.729508125s peak=1.9 GiB in_use=378.7 MiB ==
```

It saves 2.6GiB when proving the claims.